Add service to keep countdown when ui is backgrounded
This change introduced a few new components to make countdown work in
the background.

- EmergencyActionForegroundService: posts a notification with remaining
  seconds.
  * It also starts/stops sound associated with the notification
- EmergencyActionBroadcastReceiver: handles 2 action
  * calling emergency number and
  * cancelling the count down notification

When UI is closed, it start a foreground service to take over the count
down.

Then the service sends an notification immediately, and schedules a
broadcast when countdown reaches 0. This broadcast will trigger
emergency calling.

The notification has a cancel button, which also is a broadcast to
cancel ongoing foreground service.

When foreground service stops, the vibration, sound and notification all
stops with it.

Bug: 172075832
Test: manual
Change-Id: I04f33848b9532534f1551f20320683331207aa7e